+++ deserializers/guest.js
@@ -1,8 +1,0 @@
-const JSONAPIDeserializer = require('jsonapi-serializer').Deserializer;
-
-module.exports = new JSONAPIDeserializer({
-  keyForAttribute: 'camelCase',
-  projects: {
-    valueForRelationship: (relationship, included) => included,
-  },
-});