+++ services/caster.js
@@ -1,5 +1,0 @@
-function Caster() {
-  this.toBoolean = value => ['true', true, '1', 1].includes(value);
-}
-
-module.exports = Caster;