+++ templates/app/middlewares/welcome.js
@@ -1,0 +1,7 @@
+const path = require('path');
+
+module.exports = function (app) {
+  app.get('/', (req, res) => {
+    res.sendFile(path.join(__dirname, '../views/index.html'));
+  });
+};